requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.8.2 
aiohttp==3.10.5 
sympy==1.12
pytest==7.4.0
pytest-asyncio==0.21.0
streamlit==1.39.0
python-dotenv==1.0.0